requirements.txt
PyMuPDF==1.26.3
pdfplumber==0.10.3
pytesseract==0.3.10
opencv-python==4.9.0.80
pandas==2.1.4
numpy==1.26.4
Pillow==10.1.0
scikit-learn==1.4.2
